[ BE ] (perf) Students 객체의 id 타입 변경
[ BE ] (perf) Students 객체의 id 타입 변경

제 자신의 학번을 생각하며 Integer로 만들었으나, 1 부터 시작하는 것을 고려해
Integer -> Long 으로 수정하였습니다.

+++ 10 WEEK/students/src/main/java/yebin/students/Students.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class Students {
     private String name;
-    private Integer id;
+    private Long id;
     private Integer grade;
     private String major;
 
-    public Students(String name, Integer id, Integer grade, String major){
+    public Students(String name, Long id, Integer grade, String major){
         this.name = name;
         this.id = id;
         this.grade = grade;
